+++ src/main/java/com/johnbryce/talent/hackathon/repository/SubmissionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.johnbryce.talent.hackathon.models.Submission;
 
-public interface SubmissionRepository extends JpaRepository<Submission, Long> {
+public interface SubmissionRepository extends JpaRepository<Submission, Integer> {
 
 }
